refactor: extracts next/prev keyboard pattern
